Resolve compiler warnings (with and without C++11 enabled) and cleaned up the GCC diagnostic pragmas in the process. * Photivo should still compile with pre-4.6 GCC, but "unknown pragma" warnings may appear because of "pragma GCC diagnostic push/pull" (introduced in 4.6). * Photivo default branch does not yet use C++11. Accordingly the .pro files do not enable C++11 for GCC. However that will change soonish.

+++ Sources/ptAdobeTable.h
@@ -26,11 +26,11 @@
 
 struct ptAdobeTableItem {
   const char *Identification;
-  short       Blackpoint;
-  short       Whitepoint;
+  int       Blackpoint;
+  int       Whitepoint;
   // TODO Double check. But according to my (jpta) analysis
   // those are in fact XYZ->CAM matrices. And not the other way around.
-  short       XYZToCam[12];
+  int       XYZToCam[12];
 };
 
 static const ptAdobeTableItem AdobeTable[] = {
